+++ packages/core/src/method.ts
@@ -1,3 +1,3 @@
-export const SUPPORTED_METHODS = ["GET", "POST", "PUT", "PATCH", "DELETE"] as const;
+export const SUPPORTED_METHODS = ["GET", "POST", "PUT", "PATCH", "DELETE", "OPTIONS"] as const;
 
 export type RouteMethod = typeof SUPPORTED_METHODS[number];